Remove/cleanup request context helpers (#525)
* Remove context helpers in context.go
* Update request context funcs to take concrete types
* Move TestNativeContextMiddleware to mux_test.go
* Clarify KeepContext Go 1.7+ comment

Mux doesn't build on Go < 1.7 so the comment doesn't really need to
clarify anymore.
